requirements.txt
numpy==1.*
pandas
Pillow
torch==1.*
transformers==4.17.*
tqdm==4.*
sklearn
